requirements.txt
openai
python-dotenv
streamlit
langchain
langchain_core
langchainhub
langchain_openai
pydantic